refactor: add repositories folder in database context
