+++ api/src/infra/database/repositories/payable.repository.ts
@@ -1,6 +1,6 @@
 import { IPayable } from '../../../modules/payable/interfaces/payable.interface';
 import { IPayableRepository } from '../../../modules/payable/interfaces/payable.repository.interface';
-import { PrismaService } from './prisma.service';
+import { PrismaService } from '../prisma/prisma.service';
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
