+++ api/src/modules/users/test/users.service.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { UsersService } from '../users.service';
 import { mock } from 'jest-mock-extended';
 import { IUserRepository } from '../interface/users-repository.interface';
-import { UsersRepository } from '../../../infra/database/prisma/users.repository';
+import { UsersRepository } from '../../../infra/database/repositories/users.repository';
 
 describe('UsersService', () => {
   let service: UsersService;
